requirements.txt
chardet==5.2.0
dj_database_url==2.3.0
Django==5.1.7
djangorestframework==3.15.2
lxml==5.3.1
PyPDF2==3.0.1
python-dotenv==1.1.0
psycopg2-binary==2.9.10
gunicorn==23.0.0